requirements.txt
openai>=1.0.0
python-dotenv==1.0.1
google-api-python-client>=2.0.0
beautifulsoup4>=4.10.0
requests>=2.25.0
lxml>=4.6.0

# Streamlit and visualization
streamlit==1.39.0
streamlit_folium>=0.10.0
plotly==5.24.1
folium>=0.12.1

# PDF and parsing
pdfplumber==0.11.4
PyPDF2==3.0.1

# Data and utilities
pandas==2.2.3
tqdm==4.66.5
asyncio-extras==1.3.2

# LangChain ecosystem
langgraph==0.2.50
langchain==0.3.7
langchain-community==0.3.5
langchain-groq==0.2.1
pydantic==2.9.2
